Enact changes from code review

+++ server.ts
@@ -49,8 +49,8 @@ app.post<{}, {}, IResource>("/resources", async (req, res) => {
   }
 });
 
-app.post<{res_id: string}, {}, {comment_body: string, user_id: number}>("/resources/:res_id/comments", async (req, res) => {
-  const res_id = parseInt(req.params.res_id);
+app.post<{res_id: number}, {}, {comment_body: string, user_id: number}>("/resources/:res_id/comments", async (req, res) => {
+  const {res_id} = req.params;
   const {comment_body, user_id} = req.body;
   try {
     const dbResponse = await client.query(`INSERT INTO comments (comment_body, user_id, resource_id) VALUES ($1, $2, $3) RETURNING *`, [comment_body, user_id, res_id]);
@@ -62,20 +62,16 @@ app.post<{res_id: string}, {}, {comment_body: string, user_id: number}>("/resour
 });
 
 // See if it is possible to return from SQL statement
-app.post<{res_id: string}, {}, {user_id: number, like_or_dislike: "like" | "dislike"}>("/resources/:res_id/likes", async (req, res) => {
-  const res_id = parseInt(req.params.res_id);
+app.post<{res_id: number}, {}, {user_id: number, like_or_dislike: "like" | "dislike"}>("/resources/:res_id/likes", async (req, res) => {
+  const {res_id}= req.params;
   const {user_id, like_or_dislike} = req.body;
   const like_boolean = like_or_dislike === "like" ? true : false;
   try {
-    const dbResponse = await client.query(`do 
-    $do$
-    begin
-    if exists (select * from likes where user_id = $1 and resource_id = $2) then update likes set liked = $3 where user_id = $1 and resource_id = $2;
-    else insert into likes values ($1, $2, $3);
-    end if;
-    end 
-    $do$`, [user_id, res_id, like_boolean]);
-    res.status(200);
+    const dbResponse = await client.query(`
+    INSERT INTO likes VALUES ($1,$2,$3) 
+    ON CONFLICT (user_id,resource_id) 
+    DO UPDATE SET liked = $3 RETURNING *; `, [user_id, res_id, like_boolean]);
+    res.status(200).json(dbResponse);
   } catch (error) {
     console.error(error);
     res.status(400).json({status: error});
@@ -93,8 +89,8 @@ app.post<{}, {}, {tag_name: string}>("/tags", async (req, res) => {
   }
 });
 
-app.post<{user_id: string}, {}, {resource_id: number}>("/users/:user_id/study_list", async (req, res) => {
-  const user_id = parseInt(req.params.user_id);
+app.post<{user_id: number}, {}, {resource_id: number}>("/users/:user_id/study_list", async (req, res) => {
+const {user_id} = req.params;
   const {resource_id} = req.body;
   try {
     const dbResponse = await client.query(`INSERT INTO study_list (user_id, resource_id) VALUES ($1, $2) RETURNING *`, [user_id, resource_id]);
@@ -109,19 +105,15 @@ app.post<{user_id: string}, {}, {resource_id: number}>("/users/:user_id/study_li
 app.get("/resources", async (req, res) => {
   try {
     const response = await client.query("select * from resources order by time_date desc");
-    if (response.rowCount > 0) {
-      res.status(200).json(response.rows);
-    } else {
-      res.status(404).json({message: "Could not find any rows"})
-    }
+    res.status(200).json(response.rows);
   } catch (error) {
     console.error(error);
-    res.status(400).json({message: "Ooops"});
+    res.status(400).json(error);
   }
 })
 
 // GET /resources/:res-id //get a given resource
-app.get<{res_id: number}>("/resources/:res_id", async (req, res) => { //add type later
+app.get<{res_id: number}>("/resources/:res_id", async (req, res) => { 
   const {res_id} = req.params;
   try {
     const response = await client.query("select * from resources where resource_id = $1", [res_id]);
@@ -132,7 +124,7 @@ app.get<{res_id: number}>("/resources/:res_id", async (req, res) => { //add type
     }
   } catch (error) {
     console.error(error);
-    res.status(400).json({message: "Ooops"});
+    res.status(400).json(error);
   }
 })
 
@@ -141,14 +133,10 @@ app.get<{res_id: number}>("/resources/:res_id/comments", async (req, res) => {
   const {res_id} = req.params
   try {
     const response = await client.query("select * from comments where resource_id = $1", [res_id]);
-    if (response.rowCount > 0) {
-      res.status(200).json(response.rows);
-    } else {
-      res.status(404).json({message: "Could not find any rows"})
-    }
+    res.status(200).json(response.rows);
   } catch (error) {
     console.error(error);
-    res.status(400).json({message: "Ooops"});
+    res.status(400).json(error);
   }
 })
 
@@ -156,14 +144,10 @@ app.get<{res_id: number}>("/resources/:res_id/comments", async (req, res) => {
 app.get("/tags", async (req, res) => {
   try {
     const response = await client.query("select * from tags");
-    if (response.rowCount > 0) {
-      res.status(200).json(response.rows);
-    } else {
-      res.status(404).json({message: "Could not find any rows"})
-    }
+    res.status(200).json(response.rows);
   } catch(error) {
     console.error(error);
-    res.status(400).json({message: "Ooops"});
+    res.status(400).json(error);
   }
 })
 
@@ -171,30 +155,22 @@ app.get("/tags", async (req, res) => {
 app.get("/users", async (req, res) => {
   try {
     const response = await client.query("select * from users order by name asc");
-    if (response.rowCount > 0) {
-      res.status(200).json(response.rows);
-    } else {
-      res.status(404).json({message: "Could not find any rows"})
-    }
+    res.status(200).json(response.rows);
   } catch(error) {
     console.error(error);
-    res.status(400).json({message: "Ooops"});
+    res.status(400).json(error);
   }
-})
+});
 
 // GET /users/:user-id/study-list //get user's study list
 app.get<{user_id: number}>("/users/:user_id/study-list", async (req, res) => {
   const {user_id} = req.params;
   try {
-    const response = await client.query("select * from study_list where user_id = $1", [user_id]);
-    if (response.rowCount > 0) {
-      res.status(200).json(response.rows);
-    } else {
-      res.status(404).json({message: "Could not find any rows"})
-    }
+    const response = await client.query("select * from study_list join resources on study_list.resource_id = resources.resource_id where study_list.user_id = $1", [user_id]);
+    res.status(200).json(response.rows);
   } catch (error) {
     console.error(error);
-    res.status(400).json({message: "Surprise!"});
+    res.status(400).json(error);
   }
 })
 
@@ -206,11 +182,11 @@ app.delete<{res_id: number}>("/resources/:res_id", async (req, res) => {
     if (response.rowCount === 1) {
       res.status(200).json({status: "success", message: `Deleted resource ${res_id}`})
     } else {
-      res.status(400).json({message: "This could be bad!"});
+      res.status(400).json({message: "Did not delete exactly one resource"});
     }
   } catch (error) {
     console.error(error);
-    res.status(400).json({message: "Surprise!"});
+    res.status(400).json(error);
   }
 });
 
@@ -222,65 +198,66 @@ app.delete<{comment_id: number}>("/resources/comments/:comment_id", async (req, 
     if (response.rowCount === 1) {
       res.status(200).json({status: "success", message: `Deleted comment ${comment_id}`})
     } else {
-      res.status(400).json({message: "Couldn't find comment or deleted all!"});
+      res.status(400).json({message: "Did not delete exactly one comment"});
     }
   } catch (error) {
     console.error(error);
-    res.status(400).json({message: "Surprise!"});
+    res.status(400).json(error);
   }
 }
 )
 
 // DELETE /resources/:res-id/likes //delete a like or dislike
-app.delete<{res_id: number}>("/resources/:res_id/likes", async (req, res) => {
+app.delete<{res_id: number}, {}, {user_id: number}>("/resources/:res_id/likes", async (req, res) => {
   const {res_id} = req.params;
+  const {user_id} = req.body;
   try {
-    const response = await client.query("delete from likes where resource_id = $1 returning *", [res_id]);
+    const response = await client.query("delete from likes where resource_id = $1 and user_id=$2 returning *", [res_id, user_id]);
     if (response.rowCount === 1) {
       res.status(200).json({status: "success", message: `Deleted your like/dislike from ${res_id}`})
     } else {
-      res.status(400).json({message: "Couldn't find your like/dislike or deleted all!"});
+      res.status(400).json({message: "Did not delete exactly one like/dislike"});
     }
   } catch (error) {
     console.error(error);
-    res.status(400).json({message: "Surprise!"});
+    res.status(400).json(error);
   }
 }
 )
 
 // DELETE /tags //delete a tag from the database
-app.delete("/tags", async (req, res) => {
+app.delete<{}, {}, {tag_name: string}>("/tags", async (req, res) => {
   const {tag_name} = req.body;
   try {
     const response = await client.query("delete from tags where tag_name = $1 returning *", [tag_name]);
     if (response.rowCount === 1) {
-      res.status(200).json({status: "success", message: `Deleted the tag ${tag_name}`})
+      res.status(200).json({status: "success", message: `Deleted the tag ${tag_name}`});
     } else {
-      res.status(400).json({message: "Something went wrong"});
+      res.status(400).json({message: "Did not delete exactly one tag"});
     }
   } catch (error) {
     console.error(error);
-    res.status(400).json({message: "Surprise!"});
+    res.status(400).json(error);
   }
 })
 
 // DELETE /users/:user-id/study-list //delete resource from user's study list
-app.delete<{user_id: number}>("/users/:user_id/study-list", async (req, res) => {
+app.delete<{user_id: number}, {}, {resource_id: number}>("/users/:user_id/study-list", async (req, res) => {
   const {user_id} = req.params;
   const {resource_id} = req.body;
   try {
     const response = await client.query("delete from study_list where user_id = $1 and resource_id = $2 returning *", 
-      [user_id, parseInt(resource_id)]);
+      [user_id, resource_id]);
     if (response.rowCount === 1) {
       res.status(200).json({status: "success", message: `Deleted resource ${resource_id} from your study-list`})
     } else {
-      res.status(400).json({message: "Something went wrong"});
+      res.status(400).json({message: "Did not delete exactly one resource from the study list"});
     }
   } catch (error) {
     console.error(error);
-    res.status(400).json({message: "Surprise!"});
+    res.status(400).json(error);
   }
-})
+});
 
 
 //Start the server on the given port
